Moved component files into new folders
