+++ src/pages/results/components/ResultsDetailed.tsx
